Add files via upload
This was a very simple code that I made to fill a data conversion need. 

Data came in from a tool in this format: POLYGON ((-80.6166339 25.2756544, -65.8070636 18.4287094, -64.7523761 32.0929608, -80.6166339 25.2756544))

And we needed to work with it in this JSON format: {"geometry":{"type":"Polygon","coordinates":[[[-80.6166339,25.2756544],[-65.8070636,18.4287094],[-64.7523761,32.0929608],[-80.6166339,25.2756544]]]}}

It does some hacky string manipulation, but it works and it is used to process hundreds of records like this every month.

The data here is kept very high level to show the general way it is done, but for company privacy does not give too much detail.